Extract ConfirmSwapModalSetup from Swap component
